buyHouse FIXED
buy house from the front end is fixed now and transaction works. the following errors were in code:
1. i was doubling multiplication on the front end (it went toWei twice) which was creating astronomically large number and the actual price entered into the contract was 3X toWei which was sending an obscene number to contract. On the contract level, i was also doing the same calculation, which together left no option to the code but to error out (I too would be confused!).

Options left were to either change the multiplication functionality on blockchain and bring it to the front end (this would make the utilization of blockchain sort of useless) or to change how the front end operates and how house price is actually entered into smart contract. I opted to change front end and leave contract in tact.

This required reworking the front end to change price when selling from "toWei" to "BN" and then do front end multiplication using JS to accurately display data/prices.

It also meant that when buying house,  we'd use the price as is and push that in value to contract.

2. Final error occurred in the smart contract itself. Under buy function I used require(token.universalTransfer...) and that kept throwing an error at this point. Removing the "require" statement made the transaction go through, transferring correct amount of ETH as well as NFT to appropriate parties.

additional tests needed but it seems the original error of not being able to transact has been resolved.